requirements.txt
click==7.1.1
Flask==1.1.2
Flask-Cors==3.0.8
Flask-HTTPAuth==4.0.0
flask-marshmallow==0.12.0
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.6.0
marshmallow-sqlalchemy==0.23.0
numpy==1.18.3
pandas==1.0.3
Pillow==7.1.2
PyMySQL==0.9.3
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
qrcode==6.1
six==1.14.0
SQLAlchemy==1.3.16
Werkzeug==1.0.1
